regroup test and benchmark utils.
