+++ test/utils.go
@@ -10,6 +10,22 @@ import (
 	ma "github.com/multiformats/go-multiaddr"
 )
 
+func peerId(ids string) peer.ID {
+	id, err := peer.IDB58Decode(ids)
+	if err != nil {
+		panic(err)
+	}
+	return id
+}
+
+func multiaddr(m string) ma.Multiaddr {
+	maddr, err := ma.NewMultiaddr(m)
+	if err != nil {
+		panic(err)
+	}
+	return maddr
+}
+
 type peerpair struct {
 	ID   peer.ID
 	Addr ma.Multiaddr
